Big refactor to remove kube-rs dependency
Due to some restrictions we've found by using kube-rs dependency, I
decided to write an alternative kubernetes client to be able to use the
raw k8s_openapi structs. Even if not all restrictions has been lifted
right now, I belive I will be able to do it on following versions.

One of the consequences of this change is that `k8s_openapi` structs
contains lots of Optional values and this provoked that some of the
lints are less readable right now. To aleviate this, I wrote a couple of
macros `f` (which chaign optional-as_ref fields) and `m` (which chains
optional-map fields).