Add manifest for korrecte deploy
Add an example of a Kubernetes manifest that install korrecte on a
cluster.

Right now, the docker image is located on `gniet/korrecte` and is not
uploaded from CI.

The current web interface has a single endpoint ('/evaluate') which will
lint all the available resources (which depends on the service account
permissions). By default, we create a cluster role binding which allows
listing and getting all the resources that korrecte is available to
evaluate.