Lint for similar names on ClusterRole and Role
Add a lint for similar names on Roles and ClusterRole on some of the
rule's fields.

This tries to detect typos or misconfigurations on this kind of objects.
In fact, the idea of this lint came from when I was writting the
manifest for this application.

Note also that kuberntes does not complain when using an incorrect
resource name/api groups/verbs