Extend ALB ingress controller lint
Extend lint to allow ingresses for both extenions/v1beta1 (legacy) and
networking.k8s.io/v1beta1 (new).

I thought on conveting the legacy data structures to the new one and
apply the lint always on the new one, but I finally decided to extract
the logic on an extension trait.