Move documentation to YAML file
Instead of copying a lint spec everytime, we will just add a string
identifier.

I've moved everything to a YAML file where all the lints provided by the
engine will be documented (I'm not sure about if this is a good idea or
if I should maintain it on the lint's source code).

Finally, after this changes, I've been able to dump all the
documentation to the README.