Use again kube-rs
This library has improved a lot (it was already really nice when I
stopped using it) last months and all the blockers and concerns I had
with it has been solved.

By using this, I'm able to remove the custom and ad-hoc (and bad)
implementation of an async kubernetes client.

I'm really happy of start using this crate again.